+++ framework/model/src/main/java/is/codion/framework/model/AbstractEntityTableModel.java
@@ -284,7 +284,7 @@ public abstract class AbstractEntityTableModel<E extends EntityEditModel> implem
 
 	/**
 	 * Replace the entities identified by the Entity.Key map keys with their respective value.
-	 * Note that this does not trigger {@link #filterItems()}, that must be done explicitly.
+	 * Note that this does not trigger {@link Items#filter()}, that must be done explicitly.
 	 * @param entitiesByKey the entities to replace mapped to the corresponding primary key found in this table model
 	 */
 	private void replaceEntitiesByKey(Map<Entity.Key, Entity> entitiesByKey) {